modalInfo for basket and wishlist
