+++ src/components/ProductCard.vue
@@ -28,12 +28,13 @@
 </template>
 
 <script setup>
-import { computed, nextTick, reactive, ref } from 'vue';
+import { computed } from 'vue';
 import RegularText from './text/RegularText.vue';
 import ProductPrice from './ProductPrice.vue';
 import { useWishlistStore } from '@/app/store/wishlist';
 import useBasketStore from '@/app/store/basket';
 import ModalInfo from './ui/ModalInfo.vue';
+import { useModalInfo } from '@/hooks/useModalInfo';
 const props = defineProps({
     card: {
         type: Object,
@@ -50,11 +51,8 @@ const props = defineProps({
 })
 const wishlistStore = useWishlistStore();
 const basketStore = useBasketStore();
-const modalInfo = reactive({
-    className: '',
-    text: 'Added to',
-    show: false,
-})
+const {modalInfo, setModalInfo} = useModalInfo();
+
 const isLike = computed(() => 
     wishlistStore.products.some(product => product.id === props.card.id)
 );
@@ -71,13 +69,7 @@ function toggleWishlist(card) {
         setModalInfo('Added to Wishlist', 'like');
     }
 }
-async function setModalInfo(text, className) {
-    modalInfo.text = text;
-    modalInfo.className = className;
-    modalInfo.show = false;
-    await nextTick();
-    modalInfo.show = true;
-}
+
 </script>
 
 <style scoped>
