+++ src/pages/BasketPage.vue
@@ -3,7 +3,9 @@ import useBasketStore from '@/app/store/basket';
 import EmptyList from '@/components/EmptyList.vue';
 import MiddleTitle from '@/components/text/MiddleTitle.vue';
 import RegularText from '@/components/text/RegularText.vue';
+import ModalInfo from '@/components/ui/ModalInfo.vue';
 import UserProductCard from '@/components/UserProductCard.vue';
+import { useModalInfo } from '@/hooks/useModalInfo';
 import { computed } from 'vue';
 const basketStore = useBasketStore();
 const totalPrice = computed(() => 
@@ -12,7 +14,7 @@ const totalPrice = computed(() =>
   0)
 );
 const delivery = totalPrice.value * 0.05;
-
+const {modalInfo, setModalInfo} = useModalInfo();
 </script>
 
 <template>
@@ -20,6 +22,7 @@ const delivery = totalPrice.value * 0.05;
         <div class="container">
             <div class="list">
                 <UserProductCard v-for="card in basketStore.products" :key="card.product.id"
+                    :setModalInfo="setModalInfo"
                     :card="card.product"
                     :count="card.count"
                     :isBasket="true"
@@ -48,6 +51,11 @@ const delivery = totalPrice.value * 0.05;
         </div>
     </template>
     <EmptyList v-else title="Cart"/>
+    <ModalInfo v-if="modalInfo.show" 
+        v-model:isOpen="modalInfo.show"
+        :className="modalInfo.className">
+        {{modalInfo.text}}
+    </ModalInfo>
 </template>
 
 <style scoped>
